+++ contracts/mocks/token/ERC1155721/Meta20InventoryMock.sol
@@ -1,22 +1,24 @@
 pragma solidity ^0.6.6;
 
-import "@openzeppelin/contracts/access/Ownable.sol";
+// import "@openzeppelin/contracts/access/Ownable.sol";
 import "@animoca/ethereum-contracts-core_library/contracts/utils/RichUInt256.sol";
 import "@animoca/ethereum-contracts-core_library/contracts/access/MinterRole.sol";
-import "../../../token/ERC1155721/PausableInventory.sol";
+import "@animoca/ethereum-contracts-erc20_base/contracts/metatx/ERC20Fees.sol";
+import "../../../token/ERC1155721/AssetsInventory.sol";
 
-contract PausableInventoryMock is PausableInventory, Ownable, MinterRole {
+contract Meta20InventoryMock is AssetsInventory, ERC20Fees, MinterRole {
 
     using RichUInt256 for uint256;
 
-    constructor(uint256 nfMaskLength) public PausableInventory(nfMaskLength) {}
+    constructor(uint256 nfMaskLength, address gasToken, address payoutWallet
+    ) public AssetsInventory(nfMaskLength) ERC20Fees(gasToken, payoutWallet) {}
 
     /**
      * @dev Gets the token name
      * @return string representing the token name
      */
     function name() external override view returns(string memory) {
-        return "PausableInventoryMock";
+        return "MetaInventoryMock";
     }
 
     /**
@@ -24,14 +26,14 @@ contract PausableInventoryMock is PausableInventory, Ownable, MinterRole {
      * @return string representing the token symbol
      */
     function symbol() external override view returns(string memory) {
-        return "PIM";
+        return "MIM";
     }
 
     /**
      * @dev This function creates the collection id.
      * @param collectionId collection identifier
      */
-    function createCollection(uint256 collectionId) onlyOwner external {
+    function createCollection(uint256 collectionId) external onlyOwner {
         require(!isNFT(collectionId));
         emit URI(_fullUriFromId(collectionId), collectionId);
     }
@@ -65,7 +67,7 @@ contract PausableInventoryMock is PausableInventory, Ownable, MinterRole {
      * @param to address recipient that will own the minted tokens
      * @param tokenId uint256 ID of the token to be minted
      */
-    function mintNonFungible(address to, uint256 tokenId) onlyMinter external {
+    function mintNonFungible(address to, uint256 tokenId) external onlyMinter {
         require(isNFT(tokenId));
         _mintNonFungible(to, tokenId);
     }
@@ -103,7 +105,7 @@ contract PausableInventoryMock is PausableInventory, Ownable, MinterRole {
      * @param collection uint256 ID of the fungible collection to be minted
      * @param value uint256 amount to mint
      */
-    function mintFungible(address to, uint256 collection, uint256 value) onlyMinter external {
+    function mintFungible(address to, uint256 collection, uint256 value) external onlyMinter {
         require(isFungible(collection));
         _mintFungible(to, collection, value);
     }
@@ -158,5 +160,25 @@ contract PausableInventoryMock is PausableInventory, Ownable, MinterRole {
      */
     function _fullUriFromId(uint256 id) internal pure returns (string memory) {
         return string(abi.encodePacked("https://prefix/json/", id.toString()));
+    }
+
+    /**
+     * @dev Replacement for msg.sender. Returns the actual sender of a transaction: msg.sender for regular transactions,
+     * and the end-user for GSN relayed calls (where msg.sender is actually `RelayHub`).
+     *
+     * IMPORTANT: Contracts derived from {GSNRecipient} should never use `msg.sender`, and use {_msgSender} instead.
+     */
+    function _msgSender() internal virtual override(Context, ERC20Fees) view returns (address payable) {
+        return super._msgSender();
+    }
+
+    /**
+     * @dev Replacement for msg.data. Returns the actual calldata of a transaction: msg.data for regular transactions,
+     * and a reduced version for GSN relayed calls (where msg.data contains additional information).
+     *
+     * IMPORTANT: Contracts derived from {GSNRecipient} should never use `msg.data`, and use {_msgData} instead.
+     */
+    function _msgData() internal virtual override(Context, ERC20Fees) view returns (bytes memory) {
+        return super._msgData();
     }
 }