+++ contracts/token/ERC1155721/AssetsInventory.sol
@@ -1,4 +1,4 @@
-pragma solidity = 0.6.2;
+pragma solidity ^0.6.6;
 
 import "@openzeppelin/contracts/math/SafeMath.sol";
 import "@openzeppelin/contracts/utils/Address.sol";
@@ -27,7 +27,7 @@ import "./../ERC1155/IERC1155TokenReceiver.sol";
     If non-fungible bitmask length == 1, there is one Non-Fungible Collection represented by the most significant bit set to 1 and other bits set to 0.
     If non-fungible bitmask length > 1, there are multiple Non-Fungible Collections.
  */
-abstract contract AssetsInventory is IERC165, IERC721, IERC1155, IERC1155AssetCollections, IERC721Metadata, IERC1155MetadataURI, Context
+abstract contract AssetsInventory is IERC165, IERC721, IERC721Metadata, IERC1155, IERC1155MetadataURI, IERC1155AssetCollections, Context
 {
     //bytes4(keccak256("onERC721Received(address,address,uint256,bytes)"))
     bytes4 constant internal ERC721_RECEIVED = 0x150b7a02;
@@ -38,7 +38,9 @@ abstract contract AssetsInventory is IERC165, IERC721, IERC1155, IERC1155AssetCo
     // bytes4(keccak256("onERC1155BatchReceived(address,address,uint256[],uint256[],bytes)"))
     bytes4 constant internal ERC1155_BATCH_RECEIVED = 0xbc197c81;
 
+    // bytes4(keccak256("supportsInterface(bytes4)"))
     bytes4 constant internal ERC165_InterfaceId = 0x01ffc9a7;
+
     bytes4 constant internal ERC1155TokenReceiver_InterfaceId = 0x4e2312e0;
 
     // id (collection) => owner => balance
@@ -66,7 +68,7 @@ abstract contract AssetsInventory is IERC165, IERC721, IERC1155, IERC1155AssetCo
      * @dev Constructor function
      * @param nfMaskLength number of bits in the Non-Fungible Collection mask
      */
-    constructor(uint256 nfMaskLength) public {
+    constructor(uint256 nfMaskLength) internal {
         require(nfMaskLength > 0 && nfMaskLength < 256);
         uint256 mask = (1 << nfMaskLength) - 1;
         mask = mask << (256 - nfMaskLength);
@@ -80,7 +82,7 @@ abstract contract AssetsInventory is IERC165, IERC721, IERC1155, IERC1155AssetCo
      * @param interfaceId interface id to query
      * @return bool if support the given interface id
      */
-    function supportsInterface(bytes4 interfaceId) public override view returns (bool) {
+    function supportsInterface(bytes4 interfaceId) external virtual override view returns (bool) {
         return (
             // ERC165 interface id
             interfaceId == 0x01ffc9a7 ||
@@ -92,20 +94,20 @@ abstract contract AssetsInventory is IERC165, IERC721, IERC1155, IERC1155AssetCo
             interfaceId == 0x4f558e79 ||
             // ERC1155 interface id
             interfaceId == 0xd9b67a26 ||
-            // ERC1155AssetCollections interface id
-            interfaceId == 0x09ce5c46 ||
             // ERC1155MetadataURI interface id
-            interfaceId == 0x0e89341c
+            interfaceId == 0x0e89341c ||
+            // ERC1155AssetCollections interface id
+            interfaceId == 0x09ce5c46
         );
     }
 /////////////////////////////////////////// ERC721 /////////////////////////////////////////////
 
-    function balanceOf(address tokenOwner) public override view returns (uint256) {
+    function balanceOf(address tokenOwner) public virtual override view returns (uint256) {
         require(tokenOwner != address(0x0));
         return _nftBalances[tokenOwner];
     }
 
-    function ownerOf(uint256 tokenId) public override(IERC1155AssetCollections, IERC721) view returns (address) {
+    function ownerOf(uint256 tokenId) public virtual override(IERC1155AssetCollections, IERC721) view returns (address) {
         require(isNFT(tokenId));
         address tokenOwner = _owners[tokenId];
         require(tokenOwner != address(0x0));
@@ -123,7 +125,7 @@ abstract contract AssetsInventory is IERC165, IERC721, IERC1155, IERC1155AssetCo
         emit Approval(tokenOwner, to, tokenId);
     }
 
-    function getApproved(uint256 tokenId) public override view returns (address) {
+    function getApproved(uint256 tokenId) public virtual override view returns (address) {
         require(isNFT(tokenId) && exists(tokenId));
         return _tokenApprovals[tokenId];
     }
@@ -134,12 +136,12 @@ abstract contract AssetsInventory is IERC165, IERC721, IERC1155, IERC1155AssetCo
         _setApprovalForAll(sender, to, approved);
     }
 
-    function _setApprovalForAll(address sender, address operator, bool approved) internal {
+    function _setApprovalForAll(address sender, address operator, bool approved) internal virtual {
         _operatorApprovals[sender][operator] = approved;
         emit ApprovalForAll(sender, operator, approved);
     }
 
-    function isApprovedForAll(address tokenOwner, address operator) public override(IERC1155, IERC721) view returns (bool) {
+    function isApprovedForAll(address tokenOwner, address operator) public virtual override(IERC1155, IERC721) view returns (bool) {
         return _operatorApprovals[tokenOwner][operator];
     }
 
@@ -216,7 +218,7 @@ abstract contract AssetsInventory is IERC165, IERC721, IERC1155, IERC1155AssetCo
         require(_checkERC1155AndCallSafeBatchTransfer(sender, from, to, ids, values, data));
     }
 
-    function balanceOf(address tokenOwner, uint256 id) public override view returns (uint256) {
+    function balanceOf(address tokenOwner, uint256 id) public virtual override view returns (uint256) {
         require(tokenOwner != address(0x0));
 
         if (isNFT(id)) {
@@ -229,7 +231,7 @@ abstract contract AssetsInventory is IERC165, IERC721, IERC1155, IERC1155AssetCo
     function balanceOfBatch(
         address[] memory tokenOwners,
         uint256[] memory ids
-    ) public override view returns (uint256[] memory)
+    ) public virtual override view returns (uint256[] memory)
     {
         require(tokenOwners.length == ids.length);
 
@@ -252,7 +254,7 @@ abstract contract AssetsInventory is IERC165, IERC721, IERC1155, IERC1155AssetCo
 
 /////////////////////////////////////////// ERC1155AssetCollections /////////////////////////////////////////////
 
-    function collectionOf(uint256 id) public override view returns (uint256) {
+    function collectionOf(uint256 id) public virtual override view returns (uint256) {
         require(isNFT(id));
         return id & NF_COLLECTION_MASK;
     }
@@ -262,7 +264,7 @@ abstract contract AssetsInventory is IERC165, IERC721, IERC1155, IERC1155AssetCo
         @param id The ID to query
         @return bool whether the given id is fungible
      */
-    function isFungible(uint256 id) public override view returns (bool) {
+    function isFungible(uint256 id) public virtual override view returns (bool) {
         return id & (NF_BIT_MASK) == 0;
     }
 
@@ -271,7 +273,7 @@ abstract contract AssetsInventory is IERC165, IERC721, IERC1155, IERC1155AssetCo
         @param id The ID to query
         @return bool whether the given id is non-fungible token
      */
-    function isNFT(uint256 id) internal view returns (bool) {
+    function isNFT(uint256 id) internal virtual view returns (bool) {
         // A base type has the NF bit and an index
         return (id & (NF_BIT_MASK) != 0) && (id & (~NF_COLLECTION_MASK) != 0);
     }
@@ -281,7 +283,7 @@ abstract contract AssetsInventory is IERC165, IERC721, IERC1155, IERC1155AssetCo
      * @param id uint256 ID of the NFT
      * @return whether the NFT belongs to someone
      */
-    function exists(uint256 id) public view returns (bool) {
+    function exists(uint256 id) public virtual view returns (bool) {
         address tokenOwner = _owners[id];
         return tokenOwner != address(0x0);
     }
@@ -319,7 +321,7 @@ abstract contract AssetsInventory is IERC165, IERC721, IERC1155, IERC1155AssetCo
      * @param id uint256 ID of the token to be transferred
      * @param operatable bool to indicate whether the msg sender is operator
     */
-    function _transferNonFungible(address from, address to, uint256 id, bool operatable) internal {
+    function _transferNonFungible(address from, address to, uint256 id, bool operatable) internal virtual {
         require(from == _owners[id]);
 
         address sender = _msgSender();
@@ -349,7 +351,7 @@ abstract contract AssetsInventory is IERC165, IERC721, IERC1155, IERC1155AssetCo
      * @param collectionId uint256 ID of the fungible token to be transferred
      * @param value uint256 transfer amount
      */
-    function _transferFungible(address from, address to, uint256 collectionId, uint256 value) internal {
+    function _transferFungible(address from, address to, uint256 collectionId, uint256 value) internal virtual {
         _balances[collectionId][from] = SafeMath.sub(_balances[collectionId][from], value);
 
         if (to != address(0x0)) {