+++ contracts/token/ERC1155721/NonBurnablePausableInventory.sol
@@ -1,4 +1,4 @@
-pragma solidity = 0.6.2;
+pragma solidity ^0.6.6;
 
 import "./PausableInventory.sol";
 
@@ -8,7 +8,7 @@ import "./PausableInventory.sol";
 abstract contract NonBurnablePausableInventory is PausableInventory
 {
 
-    constructor(uint256 nfMaskLength) public PausableInventory(nfMaskLength)  {}
+    constructor(uint256 nfMaskLength) internal PausableInventory(nfMaskLength)  {}
 
     modifier notZero(address addr) {
         require(addr != address(0x0));
@@ -21,7 +21,7 @@ abstract contract NonBurnablePausableInventory is PausableInventory
         uint256 id,
         uint256 value,
         bytes memory data
-    ) public override notZero(to) {
+    ) public virtual override notZero(to) {
         super.safeTransferFrom(from, to, id, value, data);
     }
 
@@ -31,12 +31,12 @@ abstract contract NonBurnablePausableInventory is PausableInventory
         uint256[] memory ids,
         uint256[] memory values,
         bytes memory data
-    ) public override notZero(to) {
+    ) public virtual override notZero(to) {
         super.safeBatchTransferFrom(from, to, ids, values, data);
     }
 
     function _transferFrom(address from, address to, uint256 tokenId, bytes memory data, bool safe
-    ) internal override notZero(to) {
+    ) internal virtual override notZero(to) {
         super._transferFrom(from, to, tokenId, data, safe);
     }
 }