+++ contracts/token/ERC721/IERC721.sol
@@ -1,11 +1,11 @@
-pragma solidity = 0.6.2;
+pragma solidity ^0.6.6;
 
  /**
     @title ERC721 Non-Fungible Token Standard, basic interface
     @dev See https://eips.ethereum.org/EIPS/eip-721
     Note: The ERC-165 identifier for this interface is 0x80ac58cd.
  */
-abstract contract IERC721 {
+interface IERC721 {
     event Transfer(
         address indexed _from,
         address indexed _to,
@@ -29,14 +29,14 @@ abstract contract IERC721 {
      * @param owner address to query the balance of
      * @return balance uint256 representing the amount owned by the passed address
      */
-    function balanceOf(address owner) external virtual view returns (uint256 balance);
+    function balanceOf(address owner) external view returns (uint256 balance);
 
     /**
      * @dev Gets the owner of the specified ID
      * @param tokenId uint256 ID to query the owner of
      * @return owner address currently marked as the owner of the given ID
      */
-    function ownerOf(uint256 tokenId) public virtual view returns (address owner);
+    function ownerOf(uint256 tokenId) external view returns (address owner);
 
     /**
      * @dev Approves another address to transfer the given token ID
@@ -46,7 +46,7 @@ abstract contract IERC721 {
      * @param to address to be approved for the given token ID
      * @param tokenId uint256 ID of the token to be approved
      */
-    function approve(address to, uint256 tokenId) external virtual;
+    function approve(address to, uint256 tokenId) external;
 
     /**
      * @dev Gets the approved address for a token ID, or zero if no address set
@@ -54,7 +54,7 @@ abstract contract IERC721 {
      * @param tokenId uint256 ID of the token to query the approval of
      * @return operator address currently approved for the given token ID
      */
-    function getApproved(uint256 tokenId) public virtual view returns (address operator);
+    function getApproved(uint256 tokenId) external view returns (address operator);
 
     /**
      * @dev Sets or unsets the approval of a given operator
@@ -62,7 +62,7 @@ abstract contract IERC721 {
      * @param operator operator address to set the approval
      * @param approved representing the status of the approval to be set
      */
-    function setApprovalForAll(address operator, bool approved) external virtual;
+    function setApprovalForAll(address operator, bool approved) external;
 
     /**
      * @dev Tells whether an operator is approved by a given owner
@@ -70,7 +70,7 @@ abstract contract IERC721 {
      * @param operator operator address which you want to query the approval of
      * @return bool whether the given operator is approved by the given owner
      */
-    function isApprovedForAll(address owner,address operator) external virtual view returns (bool);
+    function isApprovedForAll(address owner,address operator) external view returns (bool);
 
     /**
      * @dev Transfers the ownership of a given token ID to another address
@@ -80,7 +80,7 @@ abstract contract IERC721 {
      * @param to address to receive the ownership of the given token ID
      * @param tokenId uint256 ID of the token to be transferred
     */
-    function transferFrom(address from, address to, uint256 tokenId) external virtual;
+    function transferFrom(address from, address to, uint256 tokenId) external;
 
     /**
      * @dev Safely transfers the ownership of a given token ID to another address
@@ -94,7 +94,7 @@ abstract contract IERC721 {
      * @param to address to receive the ownership of the given token ID
      * @param tokenId uint256 ID of the token to be transferred
     */
-    function safeTransferFrom(address from, address to, uint256 tokenId) external virtual;
+    function safeTransferFrom(address from, address to, uint256 tokenId) external;
 
     /**
      * @dev Safely transfers the ownership of a given token ID to another address
@@ -109,5 +109,5 @@ abstract contract IERC721 {
      * @param tokenId uint256 ID of the token to be transferred
      * @param data bytes data to send along with a safe transfer check
      */
-    function safeTransferFrom(address from, address to, uint256 tokenId, bytes calldata data) external virtual;
+    function safeTransferFrom(address from, address to, uint256 tokenId, bytes calldata data) external;
 }