+++ contracts/token/ERC721/IERC721Receiver.sol
@@ -1,4 +1,4 @@
-pragma solidity = 0.6.2;
+pragma solidity ^0.6.6;
 
 /**
     @title ERC721 Non-Fungible Token Standard, token receiver
@@ -25,6 +25,6 @@ abstract contract IERC721Receiver {
         @param data Additional data with no specified format
         @return bytes4 `bytes4(keccak256("onERC721Received(address,address,uint256,bytes)"))`
      */
-    function onERC721Received(address operator, address from, uint256 tokenId, bytes memory data)
-    public virtual returns (bytes4);
+    function onERC721Received(address operator, address from, uint256 tokenId, bytes calldata data
+    ) external virtual returns (bytes4);
 }