+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 2.0.0 (03/05/2020)
+
+### Breaking changes
+ * Migration to `@animoca/ethereum-contracts-core_library:1.0.0` with `solc:0.6.x` and `@openzeppelin/contracts:3.x`.
+
+### New features
+ * Added `Meta20InventoryMock.sol` and related migration.
+
+
 ## 1.1.2 (23/04/2020)
 
 ### Features